Tests/LibWeb/Text/expected/abspos-relayout-direct-child-geometry.txt

live=50,0,14,16
reference=50,0,14,16
partial-layouts=0 full-layouts=1
FAIL
